refactor: split dispatch.rs (4079 lines) into sub-module directory
dispatch.rs → dispatch/
  mod.rs      (~348 lines) - HookResult, hook_entrypoint_from_stdin, state delegates
  session.rs  (~640 lines) - session lifecycle handlers (start/end/prompt/subagent)
  tools.rs    (~420 lines) - PreToolUse/PostToolUse handlers, phase detection
  events.rs   (~220 lines) - ledger event writers (commit/merge/karvi/subagent)
  helpers.rs  (~250 lines) - active plan, skill guide, auto-digest, karvi brief
  tests.rs    (~2256 lines) - all existing tests preserved unchanged

All public API unchanged (lib.rs re-exports unaffected).
Part of GH-209.
